move database-utils -> utils/database
