+++ server/__tests__/users.test.js
@@ -1,5 +1,5 @@
 const request = require('supertest');
-const dbUtils = require('../database-utils');
+const dbUtils = require('../utils/database');
 const app = require('../app');
 
 describe('test to /users', () => {
