+++ server/utils/database.js
@@ -1,4 +1,4 @@
-const models = require('./models');
+const models = require('../models');
 
 exports.clearUserDatabase = () => {
   models.User.destroy({where: {}, truncate: true}).then();
